[test]_shell_seperating_readwrite_each
