+++ src/test/java/shell/TestShell_read.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import shell.command.Document;
-import shell.command.WriteCommand;
 import shell.command.ReadCommand;
 
 import java.io.*;
@@ -15,9 +14,10 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-public class TestShell_readwrite {
+public class TestShell_read {
 
     private static final int MAX_LBA = 99;
+    private static final int MAX_BLOCK_SIZE = 100;
     private static final String TEST_VALUE = "0xABCDFFFF";
 
     @Mock
@@ -27,7 +27,6 @@ public class TestShell_readwrite {
     private PrintStream originalOut;
     private Map<Integer, String> testData; // 테스트 데이터 추가
     private Document mockDocument;
-    private WriteCommand writeCommand;
     private ReadCommand readCommand;
 
     @BeforeEach
@@ -37,7 +36,6 @@ public class TestShell_readwrite {
         System.setOut(new PrintStream(outputStream));
 
         mockDocument = mock(Document.class);
-        writeCommand = new WriteCommand(mockDocument);
         readCommand = new ReadCommand(mockDocument);
     }
 
@@ -45,34 +43,6 @@ public class TestShell_readwrite {
     void tearDown() {
         // System.out 복원
         System.setOut(originalOut);
-    }
-
-    @Test
-    void Write_정상_테스트_확인() {
-
-        String[] args = {"write", "3", TEST_VALUE};
-
-        // When
-        writeCommand.execute(args);
-
-        // Then
-        verify(mockDocument).write(3, 1, TEST_VALUE );
-
-    }
-
-    @Test
-    void Write_Input_value_오류_확인() {
-
-        String[] args = {"write", "3", "ABCDFFFF"};
-
-        // When
-        writeCommand.execute(args);
-
-        // Then
-        // Then
-        String output = outputStream.toString().trim();
-        assertTrue(output.contains("INVALID COMMAND"));
-        verify(mockDocument, never()).write(3, 1,TEST_VALUE );
     }
 
     @Test
@@ -85,6 +55,19 @@ public class TestShell_readwrite {
 
         // Then
         verify(mockDocument).read(3, 1 );
+
+    }
+
+    @Test
+    void FullRead_정상_테스트_확인() {
+        // When
+        String[] readArgs = {"fullread"};
+
+        // When
+        readCommand.execute(readArgs);
+
+        // Then
+        verify(mockDocument).read(0, MAX_BLOCK_SIZE );
 
     }
 }
