+++ src/test/java/shell/TestShell_write.java
@@ -6,7 +6,6 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import shell.command.Document;
 import shell.command.WriteCommand;
-import shell.command.ReadCommand;
 
 import java.io.*;
 import java.util.*;
@@ -15,8 +14,9 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-public class TestShell_full {
+public class TestShell_write {
 
+    private static final int MAX_LBA = 99;
     private static final int MAX_BLOCK_SIZE = 100;
     private static final String TEST_VALUE = "0xABCDFFFF";
 
@@ -28,7 +28,6 @@ public class TestShell_full {
     private Map<Integer, String> testData; // 테스트 데이터 추가
     private Document mockDocument;
     private WriteCommand writeCommand;
-    private ReadCommand readCommand;
 
     @BeforeEach
     void setUp() {
@@ -38,13 +37,40 @@ public class TestShell_full {
 
         mockDocument = mock(Document.class);
         writeCommand = new WriteCommand(mockDocument);
-        readCommand = new ReadCommand(mockDocument);
     }
 
     @AfterEach
     void tearDown() {
         // System.out 복원
         System.setOut(originalOut);
+    }
+
+    @Test
+    void Write_정상_테스트_확인() {
+
+        String[] args = {"write", "3", TEST_VALUE};
+
+        // When
+        writeCommand.execute(args);
+
+        // Then
+        verify(mockDocument).write(3, 1, TEST_VALUE );
+
+    }
+
+    @Test
+    void Write_Input_value_오류_확인() {
+
+        String[] args = {"write", "3", "ABCDFFFF"};
+
+        // When
+        writeCommand.execute(args);
+
+        // Then
+        // Then
+        String output = outputStream.toString().trim();
+        assertTrue(output.contains("INVALID COMMAND"));
+        verify(mockDocument, never()).write(3, 1,TEST_VALUE );
     }
 
     @Test
@@ -73,18 +99,5 @@ public class TestShell_full {
         String output = outputStream.toString().trim();
         assertTrue(output.contains("INVALID COMMAND"));
         verify(mockDocument, never()).write(0, MAX_BLOCK_SIZE,TEST_VALUE );
-    }
-
-    @Test
-    void FullRead_정상_테스트_확인() {
-        // When
-        String[] readArgs = {"fullread"};
-
-        // When
-        readCommand.execute(readArgs);
-
-        // Then
-        verify(mockDocument).read(0, MAX_BLOCK_SIZE );
-
     }
 }
